Updating to support Master/Agent setup.
