+++ spec/acceptance/create_spec.rb
@@ -5,14 +5,39 @@ describe 'AEM Provider', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamil
   before :context do
 
     env = <<-ENV
-      PORT=4502
-      TYPE=author
-      RUNMODES=
-      JVM_MEM_OPTS='-Xmx1024m -XX:MaxPermSize=256M'
-    ENV
+#!/bin/bash
+#
+# This script is generated by the AEM Puppet module. 
+# Do not manually edit this file.
+#
+# This file manages the configuration of the run script.
+#
+
+PORT=4502
+TYPE='author'
+RUNMODES=''
+SAMPLE_CONTENT=''
+DEBUG_PORT=
+
+
+
+JVM_MEM_OPTS='-Xmx1024m -XX:MaxPermSize=256M'
+
+DEFAULT_JVM_OPTS='-server -Djava.awt.headless=true'
+JVM_OPTS=''
+
+SLING_PROPS='conf/sling.properties'
+
+# Do we ever implement these?
+#HOST=
+#FILE_SIZE_LIMIT
+
+# This doesn't ever seem to work
+#LOG_LEVEL=
+ENV
 
     pp = <<-MANIFEST
-      File { backup => false, }
+      File { backup => false, owner => 'aem', group => 'aem' }
 
       group { 'aem' : ensure => 'present' }
 
@@ -31,6 +56,23 @@ describe 'AEM Provider', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamil
       file { '/opt/faux/crx-quickstart/bin/start-env' :
         ensure        => 'file',
         content       => "#{env}",
+        mode          => '0755',
+      }
+
+      file { '/opt/faux/crx-quickstart/bin/start' :
+        ensure        => 'file',
+        source       => "puppet:///modules/aem/start",
+        mode          => '0755',
+      }
+
+      file { '/opt/faux/crx-quickstart/bin/start.orig' :
+        ensure        => 'file',
+        content       => '',
+        mode          => '0755',
+      }
+
+      file { '/opt/faux/crx-quickstart/repository' :
+        ensure        => 'directory',
       }
 
       file { '/opt/faux/crx-quickstart/app' :
@@ -52,6 +94,10 @@ describe 'AEM Provider', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamil
         ensure      => 'absent',
         force       => 'true',
       }
+      file { '/opt/faux' :
+        ensure  => 'absent',
+        force   => 'true',
+      }
     MANIFEST
     apply_manifest(pp, :catch_failures => true)
   end
@@ -68,18 +114,8 @@ describe 'AEM Provider', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamil
       pp = <<-MANIFEST
         File { backup => false, }
 
-        aem { 'aem' :
-          ensure          => present,
+        aem::instance { 'aem' :
           source          => '/tmp/aem-quickstart.jar',
-          home            => '/opt/aem',
-          user            => 'aem',
-          group           => 'aem',
-          jvm_mem_opts    => '-Xmx2048m -XX:MaxPermSize=512M',
-          context_root    => 'contextpath',
-          runmodes        => 'development',
-          sample_content  => false,
-          debug_port      => 30303,
-          port            => 8080,
         }
 
         file { '/opt/aem' :
@@ -110,7 +146,7 @@ describe 'AEM Provider', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamil
       pp = <<-MANIFEST
         File { backup => false, }
 
-        aem { 'existing' :
+        aem::instance { 'existing' :
           ensure      => 'present',
           home        => '/opt/faux',
           source      => '/tmp/aem-quickstart.jar',
